日记打印优化
